Completed the map/reduce functionality running on SCADS.
The entire stack is now connected, and a full map/reduce job can
execute.  Namespaces can be accessed in mappers and reducers to lookup
keys.  Intermediate outputs and final results are stored into SCADS
itself.  Added some different cases to the test program to create an
index.  Combiner functions are optional.  There is a lot of room for
optimizations...