add hashing partitioning for intermediate results.
Intermediate results (mapper result) are hashed by key.  As a hack, the
mapper generates the hash value and inserts at the beginning of the full
key.  This allows for the keys to be scattered among the nodes, and
allows for partition handlers to read all the values of its hash values.